nutrition_tracker/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic-settings
google-generativeai
aiogram
streamlit
pandas
python-dotenv
pytest
pytest-asyncio
